Refactor generated test cases slightly, to speed up the tests.
The main difference is that setting up all the test tables and types are
done in one large transaction, to reduce overhead of transaction commits.
Another is that all the objects are created in a separate schema, and
we don't bother to drop them all at the end of the tests. These changes
shave off a few seconds from the test runtime on my laptop.