Use async libpq calls instead of threading for connecting to segments.
To open the libpq connections from master to all the segments, we used to
launch multiple threads, and each thread opened gp_connections_per_thread
connections, sequentially. Backend code isn't generally thread-safe, so
using threads for that was quite complicated: a lot of care had to be taken
to avoid using palloc() and elog() in the threaded code, for example, and
a lot of complexity was needed to communicate between the main thread and
the connection threads.

It was not optimal from a concurrency point of view either, as each thread
would open a bunch of connection in sequence, unless you set
gp_connections_per_thread=1, but the default was much much higher than that.
In practice, unless you change the setting, you would end up with a single
thread to open all the connections.

Rewrite the code to open the connections, using the asynchronous
PQconnectPoll() call instead of threads. Makes it a lot simpler, and now all
the connections are launched at the same time, which should reduce the
latency of the first query in a session.

The gp_connections_per_thread GUC is not needed anymore. I kept it around
as a dummy, so that old configuration files still work. I don't think we
generally guarantee that across major versions, PostgreSQL certainly
doesn't, but we actually had gp_connections_per_thread=64 in the sample
config file, and not commented out, so almost every config file out there
will contain that, so it seems better to accept it for a few releases and
not bother users or upgrade script with it just yet.