Use a temporary memory context while parsing parts of a CreateStmt.
With a heavily partitioned table, the parse analysis of each CreateStmt
consumes significant amounts of memory. By using a temporary memory
context, we can reclaim some of the garbage left behind by the parsing of
each one. In very quick testing on my laptop, this reduces the memory
consumption of parse analysis of a massively-partitioned CREATE TABLE by
about 10%, but YMMV.

To make this work, transformAttributeEncoding() and its subroutines have
to be more careful to not modify the input CreateStmt, as any new Nodes
stored in it would be allocated in the temporary context and be blown away
at the end of transformCreateStmt. That's not cool, if there are more
partitions to process, which rely on the same CreateStmt.