Backport infrastructure to restore OIDs from PostgreSQL 9.1.
I chose to take the 9.1 version of these, because there were big cleanups
between 9.0 and 9.1. In particular, it's critical for GPDB that the pg_class
OIDs to be restored, not relfilenodes.

This code evolved in the upstream over multiple commits, so it's
unfortunately difficult to pinpoint at a particular one.