Optimize test table creation.
Optimize the functions for converting a seed integer to various geometric
types, to make the test table creation go faster. I kept the formula the
same, just written in a more performant way, so the contents of the
generated table was unchanged.

Some things that were not too obvious in the old implementation:

1. Spaces in the input strings were treated as zeros (because the extracted
   char was converted to bpchar)

2. because the offset of substring is 1-based, it behaved as if there was
   an implicit space or 0-byte at the beginning of the strings, and the
   last character was ignored.

Those things are more explicit in this new implementation, although it's
not any easier to read otherwise. That's not too important, the point of
these functions is just to generate some pseudo-random data.