Fix and re-enable leastsquares test.
It was disabled back in 2011, with a commit that merely said:

> Temporarily disable leastsquares -- it's causing grief on 64 bit platforms.

I have no idea what the grief was, but presumably it's been fixed since.
And if it hasn't, we want to find out.

I made a half-baked attempt at silencing differences that were due to
source filenames and line number, in commit 0bf31cd6, but didn't notice
then that the whole test case was in fact ignored. Fix that correctly.