Change reporting of invalid command-line options be like upstream.
If the command-line contained invalid options, we tried to print the
offending option. But that was done wrong, we always printed the last
option, regardless of which option was invalid. Rather than try to fix
that, revert the code the way it is in PostgreSQL 9.1, where this code
originally comes from, i.e. just say "invalid command-line arguments"
without trying to say which one was incorrect.