Use Resource Owners to close external tables on abort.
There was a rather ugly mechanism in fileam.c, to ensure that the
utl_fclose() function was called for any still-open external table, on
abort. That relied on a global variable and assumed that there can be only
one external table open at a time, which seems like a shaky assumption.

With this patch, utl_file() is no longer guaranteed to be called on abort.
Instead, make the UTL_FILE implementations responsible for cleaning up any
low-level resources, like file descriptors or handles to external libraries
(like libcurl), themselves. Use the ResourceOwner mechanism for that.

For curl- and EXECUTE-style external tables, register a new ResourceOwner
hook to close with the libcurl handles, and launched subprocesses, on
abort. For file external tables, modify fstream.c to open the underlying
file with OpenTransientFile(), so that it gets closed automatically on
abort.

This changes the url_fopen() API, so that it never returns a NULL. Instead
of returning NULL, and reporting an error in the *response_code/string
arguments, make url_fopen() always report errors directly with ereport().
That makes the callers simpler, and allows the implementations to flexibly
report whatever details of the error they have available.