Remove unnecessary tests on gpcheckcat after ADD/DROP COLUMN.
These tests created a column-oriented appendonly table, added or dropped a
column, and then ran gpcheckcat on the cluster. That's not a very
interesting test, because if there is an inconsistency in the catalogs,
that will surely manifest itself as problems at runtime, when you try to
update the table. So running gpcheckcat doesn't add much value, and we
have plenty of tests for ADD/DROP COLUMN of AOCS tables in the main test
suite already. Furthermore, there's been plans on running gpcheckcat
at the end of the main regression tests, so that would make the gpcheckcat
element redundant too.