Don't fire INSERT or DELETE triggers on UPDATE.
With ORCA, an UPDATE is actually implemented as an DELETE + INSERT. Don't
fire INSERT or DELETE triggers in that case.

This was broken by commit 740f304eea. I wonder if we should somehow fire
UPDATE triggers in that case, but I don't see any existing code to do that
either.