Remove ResultRelInfo.ri_partSlot field.
It's redundant with ri_resultSlot. The rule was that we used ri_resultSlot
for the main target table, and ri_partSlot for partitions under the main
target table. That seems like an unnecessary distinction, let's just use
ri_resultSlot all the time.

In the loop in ExecModifyTable, make sure that es_result_relation_info
points to the parent table, between calls to ExecInsert / ExecUpdate /
ExecDelete. That apparently haven't caused problems so far, but it looks
scary not to.