Refactor creation of PartitionSelectors to use a Path node.
The old way, where inject_partition_selectors_for_join() was called with a
pre-created Plan for the inner side, but no Plan for the outer side, was a
bit weird. It was extra inconvenient for the caller, because in the
upstream, we always build the outer Plan first, so we had to reverse that
in GPDB. With the upcoming 9.1 merge, it a bigger problem, because changes
to the way Nested Loop params are passed from the outer side to the inner
side require that the outer Plan is generated before the inner Plan.

To clean that up, and to prepare for the 9.1 merge, refactor
inject_partition_selectors_for_join() so that it doesn't require a
pre-created Plan for the inner side. It now works purely on the Path
representation.