Use 'expr_list' for PARTITION BY clause, like in upstream.
The old GROUPING SETS implementation stashed some information about the
GROUP BY clauses in ParseState.p_grp_tles, which we did not want for
a window function's PARTITION BY clause. That's why we couldn't use
transformGroupClause() with the old GROUPING SETS implementation. It
wouldn't have been hard to write that differently with the old
implementation either, e.g. add a boolean flag to transformGroupClause()
to tell whether to do it or not. But now that the old GROUPING SETS
implementation is gone, we can definitely use transformGroupClause() here
like in the upstream.