Fix expected output for queries that work now.
You no longer get the "GROUPING SETS / ROLLUP / CUBE columns ambiguous"
error. That was a quick & dirty hack to stop queries that didn't work
with the old implementation. Now they work, so memorize the correct
output in the expected output.

The results for some of these queries is very long, thousands of lines.
We should perhaps trim down those test queries, but let's do that
separately.