Make the GROUPING SETS planner a little bit smarter about data distribution.
Previously, we would always create a gather motion to bring all the data
into a single node, if there were any grouping sets. This makes it a
little bit smarter, so that if every grouping set contains a particular
column, then instead of a gather motion, we create a redistribute motion
to distribute the data based on the common GROUP BY column. An important
special case of this is if you have something like "GROUP BY col1,
ROLLUP(col2)". With this commit, we now redistribute the input on "col1"
(unless it's already suitably distributed), and perform the grouping in the
segments, instead of gathering all the data to the master.

We could do better, and perform initial grouping in the segments, in
queries like "GROUP BY ROLLUP(col)", and only do the "rolling up" of the
already grouped results in the master. But that would be a much bigger
patch, I think it would require executor changes, too.