Fix restore of AO auxiliary tables in pg_upgrade.
The AO auxiliary tables have different contents on each server. Therefore,
we must dump & restore them separately on each server, even though we now
make a copy of the QD data directory to initialize the segments.

The symptom before this patch was that after upgrade, all AO tables were
empty. The current pg_upgrade test doesn't test the contents of the tables
at all, but this patch doesn't fix that.