Add debugging code, to dump the page contents to a file for comparison.
This is similar to the WAL consistency checking that exists in later
PostgreSQL versions, but tailored for this one WAL record format that
I'm changing.

With this patch, a full image of each page by the targeted WAL records
are dumped to a file, /tmp/bmdump_<segno>_insert. And at replay, the image
after replaying each WAL record is dumped to /tmp/bmdump_<segno>_redo.
You can compare the files with each other, to verify that the WAL redo
faithfully reproduces the same page contents as in normal operation.