Replace PathKey with new DistributionKey struct, in CdbPathLocus.
In PostgreSQL, a PathKey represent sort ordering, but we have been using it
in GPDB to also represent the distribution keys of hash-distribed data in
the planner. (i.e. the keys in DISTRIBUTED BY of a table, but also when
data is redistributed by some other key on the fly). That's been convenient,
and there's some precedent for that, since PostgreSQL also uses PathKey to
represent GROUP BY columns, which is quite similar to DISTRIBUTED BY.

However, there are some differences. The opfamily, strategy and nulls_first
fields in PathKey are not applicable to distribution keys. Using the same
struct to represent ordering and hash distribution is sometimes convenient,
for example when we need to test whether the sort order or grouping is
"compatible" with the distribution. But at other times, it's confusing.

Looking ahead, I'm working on a patch to generalize the "cdbhash"
mechanism, so that we'd use the normal Postgres hash opclasses for
distribution keys, instead of hard-coding support for specific datatypes.
With that, the hash operator class or family used will be an important part
of the distribution key, in addition to the datatype. Using PathKey's
pk_opfamily field for that is tempting, but I think it would create more
confusion, because normally pk_opfamily represents a btree operator family,
not hash. So repurposing that seems like a bad idea. If we need to add a
new field for that, it seems better to invent a whole new struct to
represent distribution keys, so that we have a place to add the new field
to. So that's the long-term vision where I'm going with this.