Use a single EquivalenceClass to represent distribution key.
This merges the Hashed and HashedOJ locus types. The only difference
between the two was that HashedOJ had a list of DistributionKeys for each
key column, any of which could be used to represent the distribution,
whereas a Hashed locus had only one. The HashedOJ case happened, when
a locus was built for a join, on a RestrictInfo that was not derived from
an EquivalenceClass. However, with a little bit more work in
cdbpathlocus_join(), we can always pick the EquivalenceClass of either
side of the join, to represent the distribution of the join rel. The
rules are explained in the comments in cdbpathlocus_join(). Using a
single equivalence class simplifies things everywhere that we need to
compare or work with distribution keys.

This also fixes issue https://github.com/greenplum-db/gpdb/issues/6154.
In an outer join, we must use the distribution key of the outer side
to represent the join rel, because key column from the inner side come
out as NULLs for non-matching rows, and those NULLs are scattered
across all segments, not distributed according to the distribution key
of the inner side. There was a comment in cdbpathlocus.h, which explained
that that's the way it was supposed to work: the distribution specified
by the locus was only concernted about not-NULLs. But it seems like CTAS
never got the memo.

The repro for that bug needed a CREATE TABLE AS, with a GROUP BY query on
the resulting table to show that its distribution was incorrect. However,
there were comments in cdbpathlocus_collocates_*() saying "Or would
HashedOJ ok, too?". Because of this bug, it would in fact not have been OK.
If we had accepted HashedOJ locus there, the bug would have been easier to
reproduce, as a GROUP by directly on an outer join would have manifested
the same bug. But now that HashedOJ locus type is gone, the point is moot.