Only parse enough of each COPY row, to compute which segment to send it to.
In the common case that the distribution key is one of the first columns
in a table, COPY FROM can scale a lot better, if we only parse, and call
the input functions, for the first columns in the QD, and delay parsing
the rest of the into the QEs. The overall effort is the same, but the QD
can easily become a bottleneck.

GPDB 5 did this, but we lost in the refactoring of the COPY code in the
9.1 merge. This brings it back.