Use normal hash operator classes for data distribution.
Replace the use of the built-in hashing support for built-in datatypes,
in cdbhash.c, with the normal PostgreSQL hash functions.

It is no longer enough to track which columns/expressions are used to
distribute data. You also need to know the hash function used. For that,
a new field is added to gp_distribution_policy, to record the hash
operator class used for each distribution key column. In the planner,
a new opfamily field is added to DistributionKey, to track that throughout
the planning.

Normally, if you do "CREATE TABLE ... DISTRIBUTED BY (column)", the default
hash operator class for the datatype is used. But this patch extends the
syntax so that you can specify the operator class explicitly, like
"... DISTRIBUTED BY (column opclass)". This is similar to how an operator
class can be specified for each column in CREATE INDEX.

To support upgrade, the old hash functions have been converted to special
(non-default) operator classes, named cdbhash_*_ops. For example,
if you want to use the old hash function for an integer column, you could
do "DISTRIBUTED BY (intcol cdbhash_int4_ops)". The old hard-coded whitelist
of operators that have "compatible" cdbhash functions has been replaced
by putting the compatible hash opclasses in the same operator family.
For example, all legacy integer operator classes, cdbhash_int2_ops,
cdbhash_int4_ops and cdbhash_int8_ops, are all part of the
cdbhash_integer_ops operator family).

This removes the pg_database.hashmethod field. The hash method is now
tracked on a per-table and per-column basis, using the opclasses, so it's
not needed anymore.

ORCA doesn't know about hash operator classes, or the possibility that we
might need to use a different hash function for two columns with the same
datatype. Therefore, it cannot produce correct plans for queries that mix
different distribution hash opclasses for the same datatype, in the same
query. There are checks in the DXL->Query translation, to detect that case,
and fall back to planner. As long as you stick to the default opclasses in
all tables, we let ORCA to create the plan without any regard to them, and
use the default opclasses when translating the DXL plan to a Plan tree.
We also allow the case that all tables in the query use the "legacy"
opclasses, so that ORCA can also be used after pg_upgrade. But a mix of
the two, or using any non-default opclasses, forces ORCA to fall back.

One curiosity with this is the "int2vector" and "aclitem" datatypes. They
have a hash opclass, but no b-tree operators. GPDB 4 used to allow them
as DISTRIBUTED BY columns, but we forbid that in GPDB 5, in commit
56e7c16bad. Now they are allowed again, so you can specify an int2vector
or aclitem column in DISTRIBUTED BY, but it's still pretty useless,
because the planner still can't form EquivalenceClasses on it, and will
treat it as "strewn" distribution, and won't co-locate joins.

TODOs:

* There are no legacy anyarray/anyrange opclasses. Because the cdbhash
  code for them was kind of broken.
  (see https://github.com/greenplum-db/gpdb/issues/5467).

* A few FIXME comments in the code.

Discussion: https://groups.google.com/a/greenplum.org/forum/#!topic/gpdb-dev/4fZVeOpXllQ