Add pg_upgrade check for anyarray type as a distribution key
The legacy opclasses for anyarray is actually broken, with it, identical
values might be stored in different segments, we don't want users to keep it
as a distribution key after pg_upgrade, added a check to pg_upgrade.

To be clear, we have default opclasses for both hash method and btree method,
so users still can create anyarray type as a distribution key for newly created
tables.