Handle partTabTargetlist in plan_tree_walker() and plan_tree_mutator().
I thought that after adding the fixup_subplans() pass in commit d0aea18475,
we shouldn't need the check for subplans in FindEqKey, which was added in
commit 9d63d3c184, as long as we had the walker/mutator support. But for
some reason, the regression test added in commit 9d63d3c184 passes if
the contains_subplan() check is removed, ever without the waker/mutator
support, so I'm not sure where exactly that case is still blocked.
But in any case, but let's be tidy, even if there is no ill-effect at
the moment.

The missing walker/mutator support was noted by @hsyuan back in the
comments on PR #2444, but we didn't act on it then.