Remove unused tracking of current slice ID.
I didn't understand what it was trying to accomplish, because a single
process executes exactly one slice. Aside from that, from a purely
mechanic inspection, the 'currentExecutingSliceId' field was read in
exactly one place, in ShouldPrefetchJoinQual(). And there it was passed
as argument to findSenderMotion(), which surely always returns non-NULL
when you pass the current slice ID to it, because there should be a
Motion node corresponding to every slice ID in the plan tree.