Refactor SliceTable slightly.
Store the slices as an array, instead of List, for faster and more
convenient access to individual slices.

Rename Slice struct to ExecSlice. I'm planning to change things so that we
would build some kind of a representation of the slice table already at
planning time, so let's make it clear that this is the executor's version
of the slice table.

Remove the Node "type" tag from the ExecSlice struct. We never pass them
around separately, they're always part of the array in SliceTable.